refactor(fe): move on-mounted action to contest problem component
